[simulator][machine] Move to sub package
- the machine.go is already 150+ lines with current dumbest simulation logic,
so should put it in sub package and split to several files
- Add format script in Ayi, which will call gofmt
