+++ pkg/simulator/simulator_test.go
@@ -1,10 +1,11 @@
 package simulator
 
 import (
+	"github.com/xephonhq/xephon-b/pkg/simulator/machine"
 	"testing"
 )
 
-func TestSimulatorInterface(t *testing.T){
+func TestSimulatorInterface(t *testing.T) {
 	t.Parallel()
-	var _ Simulator = (*MachineSimulator)(nil)
+	var _ Simulator = (*machine.MachineSimulator)(nil)
 }
